Add useEffect to HoursDay
