+++ client/src/Components/Statefull/Booking.jsx
@@ -2,7 +2,6 @@ import React, { useState } from "react";
 import Moment from "moment";
 import Calendar from "./Calendar/Calendar";
 import Form from './BookingInput/Form';
-// import BookingList from './BookingList';
 import HoursDay from "./HoursDay";
 import './Booking.css';
 
@@ -16,7 +15,6 @@ export default function Booking() {
             </div>
             <div className='dates'>
                 <Calendar value={value} onChange={setValue} />
-                {/* <BookingList /> */}
                 <HoursDay value={value}/>
             </div>
         </div>
